@@ -14,90 +22,122 @@
   "main": "index.js",
   "scripts": {
     "test": "npm run lint",
+<<<<<<< HEAD
     "dev": "cross-env NODE_ENV=development node ./env/server.js",
     "build": "cross-env NODE_ENV=build nuxt build --config-file env/nuxt.config.js",
     "static": "npm run build && cross-env NODE_ENV=static nuxt generate --config-file env/nuxt.config.js",
     "start": "cross-env NODE_ENV=production node ./env/server.js",
     "generate": "nuxt generate --config-file env/nuxt.config.js",
+=======
+    "dev": "nuxt",
+    "build": "nuxt build",
+    "static": "nuxt generate",
+    "static:now": "npm run static && cp now.json dist && cd dist/ && now",
+    "start": "nuxt start",
+    "start:static": "node ./env/static-server.js",
+    "production": "npm run build && npm run start",
+>>>>>>> d6b6f0ec4b1c1d27c5097950fc037de6071642ee
     "lint": "npm run lint:es && npm run lint:css",
-    "lint:es": "eslint --fix --ext .js,.vue --ext .json --ignore-path .gitignore ./",
-    "lint:css": "npx stylelint 'src/**/*.vue'",
-    "audit": "npm audit --json | npx npm-audit-html --output reports/audit.html; exit ${PIPESTATUS[0]}",
-    "browserslist": "npx browserslist",
-    "wdio": "npx wdio ./test/wdio/wdio.conf.js",
-    "wdio-local": "npx wdio ./test/wdio/wdio.conf.local.js",
+    "lint:es": "eslint --fix --ext .js,.vue,.json ./",
+    "lint:css": "stylelint \"src/**/*.vue\"",
+    "audit": "npm audit --json | npx npm-audit-html --output reports/audit.html",
+    "cypress:open": "npx cypress open --browser chrome",
+    "cypress:run": "npx cypress run --browser chrome",
     "semantic-release": "semantic-release",
-    "travis-deploy-once": "travis-deploy-once"
+    "travis-deploy-once": "travis-deploy-once",
+    "storybook:dev": "start-storybook -p 6006 -c .storybook",
+    "storybook:prod": "build-storybook -c .storybook -o storybook",
+    "snyk-protect": "snyk protect",
+    "prepare": "npm run snyk-protect",
+    "packtracker": "packtracker-upload --stats=reports/webpack/stats/modern.json --output-path=dist/_nuxt/"
   },
   "devDependencies": {
-    "@commitlint/cli": "^7.2.1",
-    "@commitlint/config-conventional": "^7.1.2",
-    "@fullhuman/postcss-purgecss": "^1.1.0",
-    "@kazupon/vue-i18n-loader": "^0.3.0",
-    "@mapbox/stylelint-processor-arbitrary-tags": "^0.2.0",
-    "babel-eslint": "^10.0.1",
-    "babel-preset-vue-app": "^2.0.0",
-    "branch-name-lint": "^1.0.5",
-    "chai": "^4.2.0",
-    "chromedriver": "^2.43.1",
-    "cross-env": "^5.2.0",
-    "eslint": "^5.8.0",
-    "eslint-plugin-chai-friendly": "^0.4.1",
-    "eslint-plugin-json": "^1.2.1",
-    "eslint-plugin-mocha": "^5.2.0",
-    "eslint-plugin-vue": "^4.7.1",
-    "express": "^4.16.4",
-    "geckodriver": "^1.14.1",
-    "httpolyglot": "^0.1.2",
-    "husky": "^1.1.3",
-    "imagemin-jpeg-recompress": "^5.1.0",
-    "imagemin-mozjpeg": "^8.0.0",
-    "imagemin-webp": "^4.1.0",
-    "imagemin-webpack-plugin": "^2.3.0",
-    "mocha": "^5.2.0",
-    "multi-loader": "^0.1.0",
-    "node-gyp": "^3.8.0",
-    "node-notifier": "^5.3.0",
-    "opn": "^5.4.0",
-    "postcss-custom-media": "^7.0.7",
-    "postcss-loader": "^3.0.0",
-    "postcss-momentum-scrolling": "^1.4.8",
-    "postcss-nested": "^4.1.0",
-    "postcss-normalize": "^7.0.1",
-    "postcss-object-fit-images": "^1.1.2",
-    "postcss-pseudo-content-insert": "^0.3.0",
-    "rucksack-css": "^1.0.2",
-    "selenium-standalone": "^6.15.3",
-    "semantic-release": "^15.10.7",
-    "semantic-release-cli": "^4.0.9",
-    "stylelint": "^9.7.1",
-    "stylelint-config-css-modules": "^1.3.0",
-    "stylelint-config-recess-order": "^2.0.0",
-    "stylelint-config-standard": "^18.2.0",
-    "svg-fill-loader": "0.0.8",
-    "svg-sprite-loader": "^4.1.3",
-    "svgo-loader": "^2.2.0",
-    "wdio-mocha-framework": "^0.6.4",
-    "wdio-sauce-service": "^0.4.12",
-    "wdio-screenshot": "^0.6.0",
-    "wdio-selenium-standalone-service": "0.0.10",
-    "wdio-spec-reporter": "^0.1.5",
-    "webdriverio": "^4.14.0",
-    "webp-loader": "^0.4.0",
-    "webpack-virtual-modules": "0.1.10"
+    "@bazzite/nuxt-optimized-images": "0.3.0",
+    "@commitlint/cli": "8.3.5",
+    "@commitlint/config-conventional": "8.3.4",
+    "@fullhuman/postcss-purgecss": "2.0.6",
+    "@mapbox/stylelint-processor-arbitrary-tags": "0.3.0",
+    "@nuxtjs/sitemap": "2.0.1",
+    "@packtracker/webpack-plugin": "2.2.0",
+    "@storybook/addon-a11y": "5.3.13",
+    "@storybook/addon-actions": "5.3.13",
+    "@storybook/addon-backgrounds": "5.3.13",
+    "@storybook/addon-knobs": "5.3.13",
+    "@storybook/addon-links": "5.3.13",
+    "@storybook/addon-notes": "5.3.13",
+    "@storybook/addon-viewport": "5.3.13",
+    "@storybook/addons": "5.3.13",
+    "@storybook/vue": "5.3.13",
+    "babel-eslint": "10.0.3",
+    "babel-loader": "8.0.6",
+    "babel-preset-vue": "2.0.2",
+    "branch-name-lint": "1.0.5",
+    "eslint": "6.8.0",
+    "eslint-plugin-chai-friendly": "0.5.0",
+    "eslint-plugin-cypress": "2.10.3",
+    "eslint-plugin-json": "2.1.0",
+    "eslint-plugin-promise": "4.2.1",
+    "eslint-plugin-security": "1.4.0",
+    "express": "4.17.1",
+    "husky": "3.1.0",
+    "imagemin-mozjpeg": "8.0.0",
+    "imagemin-webp": "5.1.0",
+    "imagemin-webpack-plugin": "2.4.2",
+    "license-webpack-plugin": "2.1.3",
+    "lint-staged": "10.0.5",
+    "lost": "8.3.1",
+    "node-gyp": "6.1.0",
+    "postcss-custom-media": "7.0.8",
+    "postcss-loader": "3.0.0",
+    "postcss-momentum-scrolling": "2.5.15",
+    "postcss-nested": "4.2.1",
+    "postcss-normalize": "8.0.1",
+    "postcss-object-fit-images": "1.1.2",
+    "postcss-url": "9.0.0",
+    "raw-loader": "4.0.0",
+    "reqlim": "0.0.0",
+    "responsive-loader": "1.2.0",
+    "rucksack-css": "1.0.2",
+    "sharp": "0.24.1",
+    "storybook-addon-vue-info": "1.4.2",
+    "stylelint": "13.2.0",
+    "stylelint-config-css-modules": "2.2.0",
+    "stylelint-config-lost": "0.0.3",
+    "stylelint-config-recess-order": "2.0.3",
+    "stylelint-config-standard": "20.0.0",
+    "svgo-loader": "2.2.1",
+    "terser": "4.6.4",
+    "vue-docgen-api": "4.11.0",
+    "vue-docgen-loader": "1.5.0",
+    "vue-storybook": "1.2.0",
+    "vue-svg-loader": "0.16.0",
+    "webpack-virtual-modules": "0.2.1"
   },
   "dependencies": {
-    "@nuxtjs/axios": "^5.3.4",
-    "@nuxtjs/pwa": "^2.6.0",
-    "default-passive-events": "^1.0.10",
-    "mime": "^2.3.1",
-    "nuxt": "^2.2.0",
-    "nuxt-i18n": "^5.4.4",
+    "@js-basics/vector": "1.41.1",
+    "@nuxtjs/axios": "5.9.5",
+    "@nuxtjs/eslint-config": "2.0.2",
+    "@nuxtjs/pwa": "3.0.0-beta.20",
+    "@nuxtjs/robots": "2.4.2",
+    "nodeify": "1.0.1",
+    "nuxt": "2.11.0",
+    "nuxt-i18n": "6.5.0",
+    "nuxt-payload-extractor": "0.0.14",
+    "nuxt-polyfill": "1.0.3",
     "object-fit-images": "github:StephanGerbeth/object-fit-images",
-    "rxjs": "^6.3.3"
+    "rxjs": "6.5.4",
+    "snyk": "1.297.1",
+    "upath": "1.2.0",
+    "verge": "1.10.2",
+    "vue-lazy-hydration": "1.0.0-beta.12"
   },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/GrabarzUndPartner/gp-vue-boilerplate.git"
-  }
+  "optionalDependencies": {
+    "picturefill": "^3.0.3",
+    "custom-event-polyfill": "^1.0.7",
+    "default-passive-events": "^1.0.10",
+    "domtokenlist-shim": "^1.2.0",
+    "intersection-observer": "^0.7.0",
+    "requestidlecallback": "^0.3.0"
+  },
+  "snyk": true
 }
