@@ -1,26 +1,52 @@
 <template>
-  <footer>
-    <!-- <nav>
-      <link-list v-bind:list="navigation"></link-list>
-    </nav> -->
-  </footer>
+  <layout-default-container
+    class="page-footer"
+    tag="footer"
+  >
+    <nav>
+      <link-list
+        :list="navigation"
+        type="page-footer"
+      />
+    </nav>
+  </layout-default-container>
 </template>
 
 <script>
+import LayoutDefaultContainer from '@/components/layouts/DefaultContainer';
 import LinkList from '~/components/molecules/LinkList';
 
 export default {
   components: {
-    'link-list': LinkList
+    LayoutDefaultContainer,
+    LinkList
   },
   props: {
+    options: {
+      type: Object,
+      default () {
+        return {};
+      }
+    },
     navigation: {
       type: Array,
-      required: true,
-      default: function () {
+      default () {
         return [];
       }
     }
   }
 };
 </script>
+
+<style lang="postcss">
+.page-footer {
+  padding: calc(20 / 375 * 100%) 0;
+  margin: 0;
+
+  @media (--xs) {
+    padding: 20px 0;
+  }
+
+  background: var(--color-light-grey);
+}
+</style>
