@@ -1,10 +1,10 @@
-module.exports = function nuxtImageMin() {
-  this.extendBuild(config => {
-    let urlLoader = config.module.rules.find(
+module.exports = function nuxtImageMin () {
+  this.extendBuild((config) => {
+    const urlLoader = config.module.rules.find(
       rule =>
         rule.loader === 'url-loader' ||
         (rule.use && rule.use.find(item => item.loader === 'url-loader'))
     );
-    urlLoader.test = /(gif)$/;
+    urlLoader.test = /(unused)$/;
   });
 };
