@@ -1,34 +1,28 @@
 const fs = require('fs');
 const path = require('path');
 const VirtualModule = require('webpack-virtual-modules');
-var chokidar = require('chokidar');
+const chokidar = require('chokidar');
 module.exports = class VirtualLang extends VirtualModule {
-  constructor() {
-    super();
-  }
-
-  apply(compiler) {
+  apply (compiler) {
     super.apply(compiler);
 
     chokidar
       .watch(
-        ['src/locales/**/*.json', '!src/locales/global', '!src/locales/*.json'],
-        {}
+        [
+          'src/locales/**/*.json', '!src/locales/global', '!src/locales/*.json'
+        ],
+        { persistent: false }
       )
       .on('all', (event, filePath) => mergeContent(compiler, this, filePath));
   }
 };
 
-function getFile(filePath) {
-  // try {
+function getFile (filePath) {
   return fs.readFileSync(path.resolve(filePath)).toString();
-  // } catch (e) {
-  //   return JSON.stringify({});
-  // }
 }
 
-function readContent(compiler, filePath) {
-  let target = getLangFilePath(filePath);
+function readContent (compiler, filePath) {
+  const target = getLangFilePath(filePath);
   if (
     compiler.inputFileSystem._virtualFiles &&
     compiler.inputFileSystem._virtualFiles[path.resolve(target)]
@@ -40,33 +34,34 @@ function readContent(compiler, filePath) {
   return {};
 }
 
-function writeContent(virtualContent, filePath, content) {
+function writeContent (virtualContent, filePath, content) {
   virtualContent.writeModule(
     getLangFilePath(filePath),
     JSON.stringify(content)
   );
 }
 
-function mergeContent(compiler, virtualContent, filePath) {
-  let lang = getLangByFilePath(filePath);
-  let result = readContent(compiler, filePath);
+function mergeContent (compiler, virtualContent, filePath) {
+  const lang = getLangByFilePath(filePath);
+  const result = readContent(compiler, filePath);
 
-  let file = getFile(path.resolve(filePath));
-  let content = Object.assign(result, { [lang]: JSON.parse(file) });
+  const file = getFile(path.resolve(filePath));
+  const content = Object.assign(result, { [lang]: JSON.parse(file) });
   writeContent(virtualContent, filePath, content);
 }
 
-function getLangFilePath(filePath) {
-  return filePath.replace(/(src\/locales\/)(\w+\/)+(\w+)(\.json)$/, function(
+function getLangFilePath (filePath) {
+  filePath = filePath.replace(/\\/g, '/');
+  return filePath.replace(/(src\/locales\/\w+\/)([\w/?]+)(\.json)$/, function (
     match,
     p1,
-    p2,
-    p3
+    p2
   ) {
-    return `${p1}${p3}.lang`;
+    return `src/locales/${p2}.lang`;
   });
 }
 
-function getLangByFilePath(filePath) {
+function getLangByFilePath (filePath) {
+  filePath = filePath.replace(/\\/g, '/');
   return /src\/locales\/(\w+)/.exec(filePath)[1];
 }
